refactor(naming): restructure StopTime components to singular/plural
Reassign component names to reflect their cardinality:
- StopTimeItem now renders a single stop time (was FlatStopTimeItem)
- StopTimesItem renders a grouped route+headsign bucket (was StopTimeItem)

FlatStopTimeItem is removed; its body is folded into StopTimeItem.
