+++ src/components/stop-time-item.stories.tsx
@@ -1,8 +1,12 @@
 import type { Meta, StoryObj } from '@storybook/react-vite';
-import type { ContextualTimetableEntry, StopServiceType } from '../types/app/transit-composed.ts';
-import type { Agency, Route } from '../types/app/transit.ts';
+import type { InfoLevel } from '../types/app/settings';
+import type {
+  ContextualTimetableEntry,
+  RouteDirection,
+  StopServiceType,
+} from '../types/app/transit-composed';
+import type { Agency, Route } from '../types/app/transit';
 import {
-  agencyLong as longAgency,
   agencyTobus as agency,
   busRoute as baseRoute,
   busRoute2 as greenRoute,
@@ -14,17 +18,20 @@ import {
   headsignShinjuku,
   headsignWaseda,
   noColorRoute,
+  routeDirectionHeadsignBoth,
+  routeDirectionHeadsignBothMatching,
+  routeDirectionHeadsignNeither,
+  routeDirectionHeadsignStopOnly,
+  routeDirectionHeadsignTripOnly,
   routeLong,
   stopHeadsignDemachiyanagi,
   stopHeadsignLong,
   stopHeadsignMusashiKoganeiSouth,
   tramRoute,
   tripHeadsignLong,
-  tripHeadsignShort,
-} from '../stories/fixtures.ts';
-import { LANG_COMPARISON_CASES } from '../stories/lang-comparison.ts';
-import { fn } from 'storybook/test';
-import { StopTimeItem } from './stop-time-item.tsx';
+} from '../stories/fixtures';
+import { LANG_COMPARISON_CASES } from '../stories/lang-comparison';
+import { StopTimeItem } from './stop-time-item';
 
 /** Create a ContextualTimetableEntry for stories. */
 function createEntry(
@@ -48,6 +55,8 @@ function createEntry(
 ): ContextualTimetableEntry {
   const depMin = overrides.departureMinutes ?? 870; // 14:30
   const totalMinutes = overrides.totalMinutes ?? 45;
+  // Default remaining mirrors the default stopIndex (3) against
+  // totalStops (15): roughly 80% of the trip still left to ride.
   const remainingMinutes = overrides.remainingMinutes ?? Math.round(totalMinutes * 0.8);
   return {
     schedule: {
@@ -56,7 +65,13 @@ function createEntry(
     },
     routeDirection: {
       route: overrides.route ?? baseRoute,
-      tripHeadsign: { name: overrides.headsign ?? '中野駅', names: {} },
+      // Default to the logical `tripHeadsignLong` fixture so detailed /
+      // verbose info levels see full i18n data without pinning the
+      // story to any specific real-world place. A string override
+      // still works for quick variants (produces an empty `names`
+      // map, single-language only).
+      tripHeadsign:
+        overrides.headsign != null ? { name: overrides.headsign, names: {} } : tripHeadsignLong,
       ...(overrides.stopHeadsign != null
         ? { stopHeadsign: { name: overrides.stopHeadsign, names: {} } }
         : {}),
@@ -81,96 +96,24 @@ function createEntry(
   };
 }
 
-/**
- * Create a ContextualTimetableEntry using the logical long-form
- * fixtures (`routeLong`, `tripHeadsignLong`, `stopHeadsignLong`).
- * Intended for place-name-independent stories that exercise
- * structural/length characteristics.
- */
-function createLogicalLongEntry(
-  overrides: Partial<{
-    arrivalMinutes: number;
-    departureMinutes: number;
-    freq: number;
-    totalMinutes: number;
-    remainingMinutes: number;
-    isTerminal: boolean;
-    isOrigin: boolean;
-    pickupType: StopServiceType;
-    dropOffType: StopServiceType;
-  }> = {},
-): ContextualTimetableEntry {
-  return {
-    ...createEntry({
-      arrivalMinutes: overrides.arrivalMinutes,
-      departureMinutes: overrides.departureMinutes,
-      totalMinutes: overrides.totalMinutes,
-      remainingMinutes: overrides.remainingMinutes,
-      freq: overrides.freq,
-      isTerminal: overrides.isTerminal,
-      isOrigin: overrides.isOrigin,
-      pickupType: overrides.pickupType,
-      dropOffType: overrides.dropOffType,
-    }),
-    routeDirection: createRouteDirection({
-      route: routeLong,
-      tripHeadsign: tripHeadsignLong,
-      stopHeadsign: stopHeadsignLong,
-    }),
-  };
-}
-
-/**
- * Create a ContextualTimetableEntry using the short-form logical
- * fixtures — short route name, short trip headsign, no stop
- * headsign. Counterpart to {@link createLogicalLongEntry}.
- */
-function createLogicalShortEntry(
-  overrides: Partial<{
-    departureMinutes: number;
-    freq: number;
-    totalMinutes: number;
-    remainingMinutes: number;
-  }> = {},
-): ContextualTimetableEntry {
-  return {
-    ...createEntry({
-      departureMinutes: overrides.departureMinutes,
-      totalMinutes: overrides.totalMinutes,
-      remainingMinutes: overrides.remainingMinutes,
-      freq: overrides.freq,
-    }),
-    routeDirection: createRouteDirection({
-      route: baseRoute,
-      tripHeadsign: tripHeadsignShort,
-    }),
-  };
-}
-
 /** now = 14:25 → 5 minutes before the default 14:30 departure. */
 const now = new Date('2026-03-30T14:25:00');
-
-/** Three entries at 14:30, 14:45, 15:00 for a typical group. */
-const threeEntries = [
-  createEntry({ departureMinutes: 870 }),
-  createEntry({ departureMinutes: 885 }),
-  createEntry({ departureMinutes: 900 }),
-];
 
 const meta = {
   title: 'StopTime/StopTimeItem',
   component: StopTimeItem,
   args: {
-    entries: threeEntries,
+    entry: createEntry(),
     now,
+    isFirst: true,
+    showRouteTypeIcon: false,
     infoLevel: 'normal',
     dataLang: ['ja'],
-    showRouteTypeIcon: false,
   },
   argTypes: {
     infoLevel: { control: 'inline-radio', options: ['simple', 'normal', 'detailed', 'verbose'] },
+    isFirst: { control: 'boolean' },
     showRouteTypeIcon: { control: 'boolean' },
-    maxDisplay: { control: { type: 'number', min: 1, max: 5, step: 1 } },
   },
   decorators: [
     (Story) => (
@@ -188,8 +131,8 @@ type Story = StoryObj<typeof meta>;
 
 export const Default: Story = {};
 
-export const SingleEntry: Story = {
-  args: { entries: [createEntry()] },
+export const NotFirst: Story = {
+  args: { isFirst: false },
 };
 
 export const WithRouteTypeIcon: Story = {
@@ -199,152 +142,178 @@ export const WithRouteTypeIcon: Story = {
 // --- Route variants ---
 
 export const GreenRoute: Story = {
-  args: {
-    entries: [
-      createEntry({ route: greenRoute, headsign: '新橋駅', departureMinutes: 870 }),
-      createEntry({ route: greenRoute, headsign: '新橋駅', departureMinutes: 890 }),
-      createEntry({ route: greenRoute, headsign: '新橋駅', departureMinutes: 910 }),
-    ],
-  },
+  args: { entry: createEntry({ route: greenRoute, headsign: '新橋駅' }) },
 };
 
 export const TramRoute: Story = {
   args: {
-    entries: [
-      createEntry({ route: tramRoute, headsign: '早稲田', departureMinutes: 870 }),
-      createEntry({ route: tramRoute, headsign: '早稲田', departureMinutes: 880 }),
-    ],
+    entry: createEntry({ route: tramRoute, headsign: '早稲田' }),
     showRouteTypeIcon: true,
   },
 };
 
 export const NoRouteColor: Story = {
-  args: {
-    entries: [
-      createEntry({ route: noColorRoute, headsign: '駅前', departureMinutes: 870 }),
-      createEntry({ route: noColorRoute, headsign: '駅前', departureMinutes: 895 }),
-    ],
-  },
+  args: { entry: createEntry({ route: noColorRoute, headsign: '駅前' }) },
 };
 
 // --- Special states ---
 
-export const TerminalEntries: Story = {
+export const Terminal: Story = {
   args: {
-    entries: [
-      createEntry({ isTerminal: true, arrivalMinutes: 870 }),
-      createEntry({ isTerminal: true, arrivalMinutes: 890 }),
-    ],
+    entry: createEntry({
+      isTerminal: true,
+      arrivalMinutes: 870,
+      departureMinutes: 870,
+    }),
   },
 };
 
-export const DropOffOnly: Story = {
-  args: {
-    entries: [
-      createEntry({ pickupType: 1, departureMinutes: 870 }),
-      createEntry({ pickupType: 1, departureMinutes: 890 }),
-    ],
-  },
+export const PickupUnavailable: Story = {
+  args: { entry: createEntry({ pickupType: 1 }) },
 };
 
 export const EmptyHeadsign: Story = {
+  args: { entry: createEntry({ headsign: '' }) },
+};
+
+// --- Headsign state axis ---
+
+/**
+ * All five logical headsign states stacked, each rendered with the
+ * same base entry. Uses the place-name-independent logical fixtures
+ * (`routeDirectionHeadsignTripOnly` etc.) so each row reads as a
+ * structural test case, not a specific trip.
+ */
+export const HeadsignPatterns: Story = {
   args: {
-    entries: [createEntry({ headsign: '' })],
+    agency,
+    entry: createEntry(),
+    infoLevel: 'detailed',
   },
-};
-
-// --- Timetable button ---
-
-export const WithTimetableButton: Story = {
-  args: { onShowTimetable: fn() },
-};
-
-// --- maxDisplay ---
-
-export const MaxDisplay1: Story = {
-  args: { maxDisplay: 1 },
-};
-
-export const MaxDisplay5: Story = {
-  args: {
-    maxDisplay: 5,
-    entries: [
-      createEntry({ departureMinutes: 870 }),
-      createEntry({ departureMinutes: 885 }),
-      createEntry({ departureMinutes: 900 }),
-      createEntry({ departureMinutes: 920 }),
-      createEntry({ departureMinutes: 940 }),
-    ],
+  render: (args) => {
+    const patterns: { label: string; routeDirection: RouteDirection }[] = [
+      { label: 'trip only (classic)', routeDirection: routeDirectionHeadsignTripOnly },
+      { label: 'stop only (keio-bus pattern)', routeDirection: routeDirectionHeadsignStopOnly },
+      { label: 'both (different)', routeDirection: routeDirectionHeadsignBoth },
+      { label: 'both (matching, redundant)', routeDirection: routeDirectionHeadsignBothMatching },
+      { label: 'neither (fallback to route name)', routeDirection: routeDirectionHeadsignNeither },
+    ];
+    return (
+      <div className="flex max-w-sm flex-col gap-4">
+        {patterns.map((p) => (
+          <div key={p.label}>
+            <div className="mb-1 text-xs font-semibold text-gray-500">{p.label}</div>
+            <div className="rounded-lg bg-[#f5f7fa] p-3 dark:bg-gray-800">
+              <StopTimeItem
+                entry={{ ...args.entry, routeDirection: p.routeDirection }}
+                now={args.now}
+                isFirst={args.isFirst}
+                showRouteTypeIcon={args.showRouteTypeIcon}
+                infoLevel={args.infoLevel}
+                dataLang={args.dataLang}
+                agency={args.agency}
+              />
+            </div>
+          </div>
+        ))}
+      </div>
+    );
   },
 };
 
 // --- Info levels ---
 
-export const Detailed: Story = {
-  args: { infoLevel: 'detailed', agency },
-};
-
-export const Verbose: Story = {
+/**
+ * All four info levels stacked so the progressive disclosure is
+ * easy to compare at a glance. Each level renders the same entry —
+ * only `infoLevel` differs — so the additive features (trip position
+ * indicator, journey time bar, verbose dump, etc.) become visible
+ * row by row.
+ *
+ * Uses the `routeDirectionHeadsignBoth` logical fixture so both the
+ * trip headsign and stop headsign are populated, exercising the
+ * full headsign-rendering path at every info level.
+ */
+export const InfoLevelComparison: Story = {
   args: {
-    infoLevel: 'verbose',
     agency,
-    entries: [
-      createEntry({ departureMinutes: 870, direction: 0 }),
-      createEntry({ departureMinutes: 885, direction: 0 }),
-      createEntry({ departureMinutes: 900, direction: 0 }),
-    ],
+    entry: {
+      ...createEntry({ direction: 0 }),
+      routeDirection: routeDirectionHeadsignBoth,
+    },
+  },
+  render: (args) => {
+    const levels: InfoLevel[] = ['simple', 'normal', 'detailed', 'verbose'];
+    return (
+      <div className="flex max-w-sm flex-col gap-4">
+        {levels.map((level) => (
+          <div key={level}>
+            <div className="mb-1 text-xs font-semibold text-gray-500">infoLevel={level}</div>
+            <div className="rounded-lg bg-[#f5f7fa] p-3 dark:bg-gray-800">
+              <StopTimeItem
+                entry={args.entry}
+                now={args.now}
+                isFirst={args.isFirst}
+                showRouteTypeIcon={args.showRouteTypeIcon}
+                infoLevel={level}
+                dataLang={args.dataLang}
+                agency={args.agency}
+              />
+            </div>
+          </div>
+        ))}
+      </div>
+    );
   },
 };
 
-// --- Multiple groups ---
+// --- Multiple items ---
 
-/** Multiple route groups as they appear in the bottom sheet. */
-export const MultipleGroups: Story = {
-  args: { entries: threeEntries },
-  render: () => {
-    const groups = [
-      {
-        entries: [
-          createEntry({ departureMinutes: 870, headsign: '中野駅' }),
-          createEntry({ departureMinutes: 885, headsign: '中野駅' }),
-          createEntry({ departureMinutes: 900, headsign: '中野駅' }),
-        ],
-      },
-      {
-        entries: [
-          createEntry({
-            route: greenRoute,
-            headsign: '新橋駅',
-            departureMinutes: 872,
-          }),
-          createEntry({
-            route: greenRoute,
-            headsign: '新橋駅',
-            departureMinutes: 892,
-          }),
-        ],
-      },
-      {
-        entries: [
-          createEntry({
-            route: tramRoute,
-            headsign: '早稲田',
-            departureMinutes: 875,
-          }),
-        ],
-      },
+/**
+ * Multiple flat items as they appear in the stop view.
+ *
+ * Mixes several row variants to exercise the full flat-list layout:
+ * a plain departure, a different-route row, a pickup-unavailable
+ * row, a tram row, and two terminal rows — one at the first
+ * position (so the terminal marker and `RelativeTime` interact) and
+ * one further down (so the terminal marker sits next to an
+ * absolute-only time). Terminal rows exercise the
+ * `departure.arrivingAbsolute` i18n key and are the easiest way to
+ * eyeball the "着" / "Arr" opt-out behaviour from Storybook.
+ */
+export const MultipleItems: Story = {
+  args: { entry: createEntry() },
+  render: (args) => {
+    const entries = [
+      createEntry({ departureMinutes: 870, headsign: '中野駅' }),
+      createEntry({ departureMinutes: 885, route: greenRoute, headsign: '新橋駅' }),
+      createEntry({
+        departureMinutes: 900,
+        headsign: '中野駅',
+        isTerminal: true,
+        arrivalMinutes: 900,
+      }),
+      createEntry({ departureMinutes: 920, pickupType: 1, headsign: '車庫前' }),
+      createEntry({ departureMinutes: 935, route: tramRoute, headsign: '早稲田' }),
+      createEntry({
+        departureMinutes: 950,
+        route: tramRoute,
+        headsign: '早稲田',
+        isTerminal: true,
+        arrivalMinutes: 950,
+      }),
     ];
     return (
       <div className="max-w-sm rounded-lg bg-[#f5f7fa] p-3 dark:bg-gray-800">
-        {groups.map((group, i) => (
+        {entries.map((entry, i) => (
           <StopTimeItem
             key={i}
-            entries={group.entries}
-            now={now}
-            infoLevel="normal"
-            dataLang={['ja']}
-            showRouteTypeIcon
-            onShowTimetable={fn()}
+            entry={entry}
+            now={args.now}
+            isFirst={args.isFirst && i === 0}
+            showRouteTypeIcon={args.showRouteTypeIcon}
+            infoLevel={args.infoLevel}
+            dataLang={args.dataLang}
           />
         ))}
       </div>
@@ -352,7 +321,64 @@ export const MultipleGroups: Story = {
   },
 };
 
-/** Long route name (no short name) — tests layout. */
+/**
+ * Multi-item list stacked across every supported language so the
+ * `departure.arrivingAbsolute` marker can be verified per locale:
+ * ja shows "着" next to terminal absolute times, en shows "Arr",
+ * and other locales follow their own key value or fall through the
+ * locale chain. Locale owners can still opt out for any language by
+ * setting the key to an empty string (the component always renders
+ * the span — visibility is driven entirely by i18n). Uses the same
+ * entry mix as {@link MultipleItems} so the comparison is
+ * row-for-row identical across languages.
+ */
+export const MultipleItemsLangComparison: Story = {
+  args: { entry: createEntry(), infoLevel: 'normal' },
+  render: (args) => {
+    const entries = [
+      createEntry({ departureMinutes: 870, headsign: '中野駅' }),
+      createEntry({ departureMinutes: 885, route: greenRoute, headsign: '新橋駅' }),
+      createEntry({
+        departureMinutes: 900,
+        headsign: '中野駅',
+        isTerminal: true,
+        arrivalMinutes: 900,
+      }),
+      createEntry({ departureMinutes: 920, pickupType: 1, headsign: '車庫前' }),
+      createEntry({
+        departureMinutes: 935,
+        route: tramRoute,
+        headsign: '早稲田',
+        isTerminal: true,
+        arrivalMinutes: 935,
+      }),
+    ];
+    return (
+      <div className="flex flex-col gap-3">
+        {LANG_COMPARISON_CASES.map(({ dataLang, label }) => (
+          <div key={label} className="space-y-1">
+            <span className="block text-[10px] text-gray-400">{label}</span>
+            <div className="max-w-sm rounded-lg bg-[#f5f7fa] p-3 dark:bg-gray-800">
+              {entries.map((entry, i) => (
+                <StopTimeItem
+                  key={i}
+                  entry={entry}
+                  now={args.now}
+                  isFirst={args.isFirst && i === 0}
+                  showRouteTypeIcon={args.showRouteTypeIcon}
+                  infoLevel={args.infoLevel}
+                  dataLang={dataLang}
+                />
+              ))}
+            </div>
+          </div>
+        ))}
+      </div>
+    );
+  },
+};
+
+/** Long route name (no short name) — tests layout wrapping. */
 const longRoute: Route = {
   ...baseRoute,
   route_id: 'toaran:SA',
@@ -368,280 +394,70 @@ const longRoute: Route = {
 /** trip empty + stop present (keio-bus pattern). */
 export const TripEmptyStopPresent: Story = {
   args: {
-    entries: [
-      createEntry({ headsign: '', stopHeadsign: '武蔵小金井駅南口' }),
-      {
-        ...createEntry({ departureMinutes: 885 }),
-        routeDirection: createRouteDirection({
-          ...createEntry({ departureMinutes: 885 }).routeDirection,
-          tripHeadsign: emptyHeadsign,
-          stopHeadsign: stopHeadsignMusashiKoganeiSouth,
-        }),
-      },
-    ],
+    entry: {
+      ...createEntry(),
+      routeDirection: createRouteDirection({
+        ...createEntry().routeDirection,
+        tripHeadsign: emptyHeadsign,
+        stopHeadsign: stopHeadsignMusashiKoganeiSouth,
+      }),
+    },
   },
 };
 
 /** stop overrides trip — stop_headsign differs from trip_headsign. */
 export const StopOverridesTrip: Story = {
   args: {
-    entries: [
-      createEntry({
-        route: longRoute,
-        headsign: '北大路BT・下鴨神社・出町柳駅',
-        stopHeadsign: '出町柳駅',
+    entry: {
+      ...createEntry(),
+      routeDirection: createRouteDirection({
+        ...createEntry().routeDirection,
+        tripHeadsign: headsignKyotoLong,
+        stopHeadsign: stopHeadsignDemachiyanagi,
       }),
-      {
-        ...createEntry({ route: longRoute, departureMinutes: 885 }),
-        routeDirection: createRouteDirection({
-          ...createEntry({ route: longRoute, departureMinutes: 885 }).routeDirection,
-          tripHeadsign: headsignKyotoLong,
-          stopHeadsign: stopHeadsignDemachiyanagi,
-        }),
-      },
-    ],
-    agency: longAgency,
+    },
   },
 };
 
-/**
- * All languages side by side using the logical long-form fixtures
- * (`routeLong`, `tripHeadsignLong`, `stopHeadsignLong`). Every
- * language row gets populated trip + stop headsigns and a verbose
- * route_long_name so wrap / truncation behaviour is exercised
- * consistently per language. A second entry uses `tripHeadsignShort`
- * to show the short-form side by side.
- */
-export const LangComparison: Story = {
-  args: {
-    agency,
-    entries: [
-      {
-        ...createEntry({ departureMinutes: 870 }),
-        routeDirection: createRouteDirection({
-          route: routeLong,
-          tripHeadsign: tripHeadsignLong,
-          stopHeadsign: stopHeadsignLong,
-        }),
-      },
-      {
-        ...createEntry({ departureMinutes: 885 }),
-        routeDirection: createRouteDirection({
-          route: routeLong,
-          tripHeadsign: tripHeadsignShort,
-        }),
-      },
-    ],
-    infoLevel: 'normal',
-  },
-  render: (args) => (
-    <div className="flex flex-col gap-3">
-      {LANG_COMPARISON_CASES.map(({ dataLang, label }) => (
-        <div key={label} className="space-y-1">
-          <span className="block text-[10px] text-gray-400">{label}</span>
-          <StopTimeItem
-            entries={args.entries}
-            now={args.now}
-            infoLevel={args.infoLevel}
-            dataLang={dataLang}
-            showRouteTypeIcon={args.showRouteTypeIcon}
-            agency={args.agency}
-            maxDisplay={args.maxDisplay}
-            onShowTimetable={args.onShowTimetable}
-          />
-        </div>
-      ))}
-    </div>
-  ),
-};
-
-/** Kitchen sink groups: various data patterns to verify layout. */
-const kitchenSinkGroups: { entries: ContextualTimetableEntry[]; agency?: Agency }[] = [
-  // Short route + short headsign
-  {
-    entries: [
-      createEntry({ departureMinutes: 870, headsign: '中野駅' }),
-      createEntry({ departureMinutes: 885, headsign: '中野駅' }),
-      createEntry({ departureMinutes: 900, headsign: '中野駅' }),
-    ],
-    agency,
-  },
-  // Short route + headsign with translations
-  {
-    entries: [
-      {
-        ...createEntry({ departureMinutes: 872 }),
-        routeDirection: createRouteDirection({
-          route: greenRoute,
-          tripHeadsign: headsignShimbashiEkimae,
-        }),
-      },
-      {
-        ...createEntry({ departureMinutes: 892 }),
-        routeDirection: createRouteDirection({
-          route: greenRoute,
-          tripHeadsign: headsignShimbashiEkimae,
-        }),
-      },
-    ],
-    agency,
-  },
-  // Long route name + headsign with translations
-  {
-    entries: [
-      {
-        ...createEntry({ route: longRoute, departureMinutes: 875 }),
-        routeDirection: createRouteDirection({
-          route: longRoute,
-          tripHeadsign: headsignMinowabashi,
-        }),
-      },
-      {
-        ...createEntry({ route: longRoute, departureMinutes: 890 }),
-        routeDirection: createRouteDirection({
-          route: longRoute,
-          tripHeadsign: headsignMinowabashi,
-        }),
-      },
-    ],
-    agency: longAgency,
-  },
-  // Long route name + different headsign with translations
-  {
-    entries: [
-      {
-        ...createEntry({ route: longRoute, departureMinutes: 880 }),
-        routeDirection: createRouteDirection({ route: longRoute, tripHeadsign: headsignWaseda }),
-      },
-    ],
-    agency: longAgency,
-  },
-  // Long route + long headsign with long translations (Kyoto-style)
-  {
-    entries: [
-      {
-        ...createEntry({ route: longRoute, departureMinutes: 882 }),
-        routeDirection: createRouteDirection({ route: longRoute, tripHeadsign: headsignKyotoLong }),
-      },
-    ],
-    agency: longAgency,
-  },
-  // Long route + terminal
-  {
-    entries: [
-      {
-        ...createEntry({
-          route: longRoute,
-          departureMinutes: 884,
-          isTerminal: true,
-          arrivalMinutes: 884,
-        }),
-        routeDirection: createRouteDirection({
-          route: longRoute,
-          tripHeadsign: headsignMinowabashi,
-        }),
-      },
-    ],
-    agency: longAgency,
-  },
-  // All elements: icon + long route + long subNames + long headsign + drop-off only + agency
-  {
-    entries: [
-      {
-        ...createEntry({
-          route: longRoute,
-          departureMinutes: 886,
-          pickupType: 1,
-        }),
-        routeDirection: createRouteDirection({ route: longRoute, tripHeadsign: headsignKyotoLong }),
-      },
-    ],
-    agency: longAgency,
-  },
-  // All elements short: icon + short route + short subNames + short headsign + drop-off only + agency
-  {
-    entries: [
-      {
-        ...createEntry({
-          departureMinutes: 888,
-          pickupType: 1,
-        }),
-        routeDirection: createRouteDirection({ route: baseRoute, tripHeadsign: headsignShinjuku }),
-      },
-    ],
-    agency,
-  },
-  // Empty headsign
-  {
-    entries: [createEntry({ departureMinutes: 905, headsign: '' })],
-    agency,
-  },
-];
-
-// --- Logical data (place-name-independent) ---
+// --- infoLevel comparison ---
 
 /**
  * Side-by-side comparison of all `infoLevel` values against the
- * logical long-form fixtures. Place-name-independent alternative to
- * {@link Detailed} / {@link Verbose} stories.
+ * logical long-form fixtures (`routeLong`, `tripHeadsignLong`,
+ * `stopHeadsignLong`). Place-name-independent — exercises the full
+ * info-level rendering range without being tied to specific
+ * real-world data.
  */
 export const LogicalLongInfoLevelComparison: Story = {
   args: {
-    // Vary per-entry attributes so the inline `TimetableEntryAttributesLabels`
-    // exercises every supported flag (terminal / origin / pickup× / dropoff×)
-    // across the displayed rows:
-    //   1st: plain                      → no labels
-    //   2nd: origin only                → 始発
-    //   3rd: kitchen sink               → 始発 + 終点 + 乗× + 降×
-    //
-    // The 3rd row deliberately combines all four flags to verify they
-    // can coexist visually within a single row without overflowing or
-    // overlapping the time text. Use `Detailed` / `Verbose` for the
-    // single-flag-per-row variant if needed.
-    entries: [
-      createLogicalLongEntry({
-        departureMinutes: 870,
-        isOrigin: false,
-        isTerminal: false,
-        pickupType: 0,
-        dropOffType: 0,
-      }),
-      createLogicalLongEntry({
-        arrivalMinutes: 855,
-        departureMinutes: 855,
-        isOrigin: true,
-        isTerminal: false,
-        pickupType: 0,
-        dropOffType: 0,
-      }),
-      createLogicalLongEntry({
-        arrivalMinutes: 900,
-        departureMinutes: 900,
-        isOrigin: true,
-        isTerminal: true,
-        pickupType: 1,
-        dropOffType: 1,
-      }),
-    ],
-    agency: longAgency,
-    onShowTimetable: fn(),
+    agency,
+    entry: {
+      ...createEntry({ departureMinutes: 870 }),
+      routeDirection: {
+        route: routeLong,
+        tripHeadsign: tripHeadsignLong,
+        stopHeadsign: stopHeadsignLong,
+        direction: 0,
+      },
+    },
+    // showRouteTypeIcon: true,
   },
   render: (args) => {
-    const levels = ['simple', 'normal', 'detailed', 'verbose'] as const;
+    const levels: InfoLevel[] = ['simple', 'normal', 'detailed', 'verbose'];
     return (
       <div className="flex flex-col gap-3">
         {levels.map((level) => (
           <div key={level} className="space-y-1">
             <span className="block text-[10px] text-gray-400">infoLevel: {level}</span>
             <StopTimeItem
-              entries={args.entries}
-              now={now}
+              entry={args.entry}
+              now={args.now}
+              isFirst={args.isFirst}
+              showRouteTypeIcon={level === 'verbose'}
               infoLevel={level}
-              dataLang={['ja']}
-              showRouteTypeIcon
+              dataLang={args.dataLang}
               agency={args.agency}
               showAgency={false}
-              onShowTimetable={args.onShowTimetable}
             />
           </div>
         ))}
@@ -650,78 +466,176 @@ export const LogicalLongInfoLevelComparison: Story = {
   },
 };
 
+// --- i18n: lang resolution ---
+
 /**
- * Logical kitchen sink — multiple groups built from logical
- * fixtures. Mirrors the structure of {@link KitchenSink} but uses
- * place-name-independent data so it tests layout characteristics
- * rather than specific names.
+ * All supported languages stacked so i18n behaviour is easy to verify
+ * at a glance. Uses the logical long-form fixtures
+ * (`tripHeadsignLong`, `stopHeadsignLong`, `routeLong`) so every
+ * language row gets a populated translation and the stop/trip
+ * headsign overflow paths are exercised per language.
  */
-export const LogicalKitchenSink: Story = {
-  args: { entries: [createLogicalShortEntry()] },
-  render: () => {
-    const groups: { entries: ContextualTimetableEntry[]; agency?: Agency }[] = [
-      // Short form — 3 entries
-      {
-        entries: [
-          createLogicalShortEntry({ departureMinutes: 870 }),
-          createLogicalShortEntry({ departureMinutes: 885 }),
-          createLogicalShortEntry({ departureMinutes: 900 }),
-        ],
-        agency,
+export const LangComparison: Story = {
+  args: {
+    agency,
+    entry: {
+      ...createEntry({ departureMinutes: 870 }),
+      routeDirection: {
+        route: routeLong,
+        tripHeadsign: tripHeadsignLong,
+        stopHeadsign: stopHeadsignLong,
+        direction: 0,
       },
-      // Long form — 3 entries (trip + stop headsign)
-      {
-        entries: [
-          createLogicalLongEntry({ departureMinutes: 872 }),
-          createLogicalLongEntry({ departureMinutes: 892 }),
-          createLogicalLongEntry({ departureMinutes: 912 }),
-        ],
-        agency: longAgency,
-      },
-      // Long form — single terminal entry
-      {
-        entries: [createLogicalLongEntry({ departureMinutes: 880, isTerminal: true })],
-        agency: longAgency,
-      },
-      // Long form — drop-off only
-      {
-        entries: [createLogicalLongEntry({ departureMinutes: 884, pickupType: 1 })],
-        agency: longAgency,
-      },
-    ];
-    return (
-      <div className="max-w-sm rounded-lg bg-[#f5f7fa] p-3 dark:bg-gray-800">
-        {groups.map((group, i) => (
-          <StopTimeItem
-            key={i}
-            entries={group.entries}
-            now={now}
-            infoLevel="detailed"
-            dataLang={['ja']}
-            showRouteTypeIcon
-            agency={group.agency}
-            onShowTimetable={fn()}
-          />
-        ))}
-      </div>
-    );
+    },
+    infoLevel: 'normal',
   },
+  render: (args) => (
+    <div className="flex flex-col gap-3">
+      {LANG_COMPARISON_CASES.map(({ dataLang, label }) => (
+        <div key={label} className="space-y-1">
+          <span className="block text-[10px] text-gray-400">{label}</span>
+          <StopTimeItem
+            entry={args.entry}
+            now={args.now}
+            isFirst={args.isFirst}
+            showRouteTypeIcon={args.showRouteTypeIcon}
+            infoLevel={args.infoLevel}
+            dataLang={dataLang}
+            agency={args.agency}
+          />
+        </div>
+      ))}
+    </div>
+  ),
 };
 
+/** Kitchen sink items: various data patterns to verify layout. */
+const kitchenSinkItems: {
+  entry: ContextualTimetableEntry;
+  agency?: Agency;
+  icon?: boolean;
+}[] = [
+  // 0分後 — まもなく, short route
+  { entry: createEntry({ departureMinutes: 865, headsign: '中野駅' }) },
+  // 1分後 — short route + headsign with translations
+  {
+    entry: {
+      ...createEntry({ departureMinutes: 866 }),
+      routeDirection: createRouteDirection({
+        route: greenRoute,
+        tripHeadsign: headsignShimbashiEkimae,
+      }),
+    },
+    agency,
+  },
+  // 1分後 — long route + short headsign
+  {
+    entry: createEntry({ route: longRoute, departureMinutes: 866, headsign: '三ノ輪橋' }),
+    icon: true,
+  },
+  // 2分後 — long route + headsign with translations
+  {
+    entry: {
+      ...createEntry({ route: longRoute, departureMinutes: 867 }),
+      routeDirection: createRouteDirection({ route: longRoute, tripHeadsign: headsignMinowabashi }),
+    },
+    icon: true,
+    agency,
+  },
+  // 3分後 — long route + headsign with translations (Waseda)
+  {
+    entry: {
+      ...createEntry({ route: longRoute, departureMinutes: 868 }),
+      routeDirection: createRouteDirection({ route: longRoute, tripHeadsign: headsignWaseda }),
+    },
+    icon: true,
+  },
+  // 3分後 — long route + long headsign (Kyoto-style)
+  {
+    entry: {
+      ...createEntry({ route: longRoute, departureMinutes: 868 }),
+      routeDirection: createRouteDirection({ route: longRoute, tripHeadsign: headsignKyotoLong }),
+    },
+    icon: true,
+    agency,
+  },
+  // 5分後 — all long + terminal
+  {
+    entry: {
+      ...createEntry({
+        route: longRoute,
+        departureMinutes: 870,
+        isTerminal: true,
+        arrivalMinutes: 870,
+      }),
+      routeDirection: createRouteDirection({ route: longRoute, tripHeadsign: headsignKyotoLong }),
+    },
+    icon: true,
+    agency,
+  },
+  // 9分後 — all long + pickup unavailable
+  {
+    entry: {
+      ...createEntry({
+        route: longRoute,
+        departureMinutes: 874,
+        pickupType: 1,
+      }),
+      routeDirection: createRouteDirection({ route: longRoute, tripHeadsign: headsignKyotoLong }),
+    },
+    icon: true,
+    agency,
+  },
+  // 10分後 — all short + terminal
+  {
+    entry: {
+      ...createEntry({
+        departureMinutes: 875,
+        isTerminal: true,
+        arrivalMinutes: 875,
+      }),
+      routeDirection: createRouteDirection({ route: baseRoute, tripHeadsign: headsignShinjuku }),
+    },
+    icon: true,
+    agency,
+  },
+  // 11分後 — long route + terminal (no relative time)
+  {
+    entry: createEntry({
+      route: longRoute,
+      departureMinutes: 876,
+      headsign: '三ノ輪橋',
+      isTerminal: true,
+      arrivalMinutes: 876,
+    }),
+    icon: true,
+  },
+  // 14分後 — pickup unavailable (no relative time)
+  { entry: createEntry({ departureMinutes: 879, headsign: '車庫前', pickupType: 1 }) },
+  // 15分後 — empty headsign
+  { entry: createEntry({ departureMinutes: 880, headsign: '' }) },
+  // 30分後
+  { entry: createEntry({ departureMinutes: 895, headsign: '中野駅' }) },
+  // 60分後
+  { entry: createEntry({ departureMinutes: 925, headsign: '中野駅' }) },
+  // 120分後
+  { entry: createEntry({ departureMinutes: 985, headsign: '中野駅' }) },
+];
+
 export const KitchenSink: Story = {
-  args: { entries: threeEntries },
-  render: () => (
+  args: { entry: createEntry(), infoLevel: 'normal' },
+  render: (args) => (
     <div className="max-w-sm rounded-lg bg-[#f5f7fa] p-3 dark:bg-gray-800">
-      {kitchenSinkGroups.map((group, i) => (
+      {kitchenSinkItems.map(({ entry, agency: a, icon }, i) => (
         <StopTimeItem
           key={i}
-          entries={group.entries}
-          now={now}
-          infoLevel="detailed"
-          dataLang={['ja']}
-          showRouteTypeIcon
-          agency={group.agency}
-          onShowTimetable={fn()}
+          entry={entry}
+          now={args.now}
+          isFirst={args.isFirst && i === 0}
+          showRouteTypeIcon={args.showRouteTypeIcon || (icon ?? false)}
+          infoLevel={args.infoLevel}
+          dataLang={args.dataLang}
+          agency={a}
         />
       ))}
     </div>
